pyproject.toml
###########
# 📜 Poetry
###########
[tool.poetry]
name = "IRIS Data Cleanser"
version = "0.1.0"
description = "IRIS Data Cleanser - ETL pipeline using Airbyte, dbt (targeting Postgres), loading into S3."
license = "UNLICENSED"
classifiers = ["Private :: Do Not Upload"]
packages = [
  { include = "IRIS Data Cleanser", from = "IRIS Data Cleanser" },
  # Add more packages here, e.g.
  # { include = "module_name", from = "IRIS Data Cleanser" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
# Everything below here is alphabetically sorted
airbyte-cdk = "<4.3.2"
boto3 = "^1.34.158"
cryptography = ">=43.0.1"
dbt-core = "<=1.5.5"
dbt-fal = "^1.5.9"
dbt-postgres = "<1.8.2"
fastapi = ">=0.109.1"
fuzzymatcher = "^0.0.6"
isolate = "^0.14.2"
ipywidgets = "^8.1.3"
joblib = "^1.4.2"
jsonschema = "<3.3.0"
jupyterlab = "<4"
jupyterlab-widgets = "^3.0.11"
numpy = "^1.26.4"
openpyxl = "^3.1.5"
pandarallel = "^1.6.5"
pandas = "<2.0"
pickleshare = "^0.7.5"
pyproj = "^3.6.1"
pyyaml = "^6.0.2"
rapidfuzz = "^3.9.6"
requests = "^2.32.3"
sqlparse = ">=0.5.0"
tqdm = "^4.66.5"
typer = "^0.9.0"
folium = "^0.17.0"
geopy = "^2.4.1"
matplotlib = "^3.9.2"
plotly = "^5.23.0"

[tool.poetry.dev-dependencies]
# Everything below here is alphabetically sorted
bandit = "^1.7.9"
detect-secrets = "^1.5.0"
ipdb = "^0.13.13"
jupyter-black = "^0.3.4"
liccheck = "^0.9.2"
pip-audit = "^2.7.3"
pre-commit = "^3.8.0"
pytest = "^7.4.4"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"
ruff = "^0.1.13"
sqlfluff = "^2.3.5"
towncrier = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

############
# ✅ Linters
############
# [tool.pylint."MESSAGES CONTROL"]
# disable = [
#   "locally-disabled",
#   "logging-fstring-interpolation",
#   "missing-class-docstring",
#   "no-else-break",
#   "no-else-raise",
#   "suppressed-message",
#   "unused-argument",
# ]


[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py312"
exclude = ["*.ipynb", "notebooks/recommendations.py"]

[tool.ruff.lint]
select = [
  "A",
  "AIR",
  "ARG",
  "B",
  "C4",
  "C90",
  "COM",
  # "CPY",
  "D",
  "DJ",
  "DTZ",
  "E",
  # "E4", "E7", "E9",
  "ERA",
  "EXE",
  "F",
  "FA",
  "FIX",
  "FLY",
  # "FURB",
  "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  # "LOG",
  "N",
  "NPY",
  "PD",
  "PERF",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "Q",
  "RET",
  "RSE",
  "RUF",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TCH",
  "TD",
  "TID",
  "TRY",
  "UP",
  "W",
]
ignore = [
  "ARG001",
  "ARG002",
  "COM812",
  "D101",
  "D202",
  "D203",
  "D205",
  "D211",
  "D213",
  "D400",
  "D400",
  "D401",
  "D404",
  "D415",
  "DTZ005",
  "E203",
  "FIX002",
  "G004",
  "INP001",
  "ISC001",
  "PD901",
  "PLR0913",
  "PLR2004",
  "PT004",
  "PTH107",
  "PTH123",
  "RET504",
  "RET506",
  "RET508",
  "RUF013",
  "T100",
  "TD002",
  "TD003",
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.sqlfluff.core]
dialect = "postgres"
templater = "jinja"
sql_file_exts = ".sql,.sql.j2,.dml,.ddl"
max_line_length = 100
exclude_rules = [
  "CP02", # CP02: Unquoted identifiers must be consistently lower case
  "L029", # L029: prevents use of 'date' as a column name
  "L032", # L032: prevents use of USING joins
  "RF05", # RF05: Do not use special characters in identifiers
  "ST06", # ST06: Select wildcards then simple targets before calculations and aggregates
]

[tool.sqlfluff.indentation]
indented_joins = false
indented_using_on = true
template_blocks_indent = false

[tool.sqlfluff.templater]
unwrap_wrapped_queries = true

[tool.sqlfluff.templater.dbt]
project_dir = "./"

[tool.sqlfluff.templater.jinja]
apply_dbt_builtins = true

[tool.sqlfluff.rules.capitalisation.keywords]
capitalisation_policy = "upper"


##############################
# ✅ liccheck
##############################
[tool.liccheck]
authorized_licenses = [
  "3-clause bsd",
  "apache 2.0",
  "apache license 2.0",
  "apache license, version 2.0",
  "apache software",
  "bsd",
  "bsd 3-clause",
  "bsd-3-clause",
  "freeware",
  "gnu library or lesser general public license (lgpl)",
  "historical permission notice and disclaimer (hpnd)",
  "isc license (iscl)",
  "mit",
  "mozilla public license 2.0 (mpl 2.0)",
  "new bsd",
  "python software foundation",
  "the unlicense (unlicense)",
]
unauthorized_licenses = [
  "gnu general public license (gpl)",
  "gnu general public license v2 (gplv2)",
  "gplv2",
]


##############
# 📣 Towncrier
##############
[tool.towncrier]
package = "IRIS Data Cleanser"
package_dir = "."
filename = "CHANGELOG.md"
directory = ".changelog/"
template = ".changelog/template.md"
title_format = "## [{version}] - {project_date}"
issue_format = "[#{issue}](https://github.com/National-Digital-Twin/IRIS-data-cleanser/issues{issue})"
start_string = "<!-- TOWNCRIER -->\n"
underlines = ["", ""]

# .added for new features
# .changed for changes in existing functionality
# .deprecated for soon-to-be removed features
# .removed for now removed features
# .fixed for any bug fixes
# .security in case of vulnerabilities
# .analysis for data analyses
# .docs for documentation improvements
# .maintenance for maintenance tasks & upgrades
[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Bug Fixes"
showcontent = true

[[tool.towncrier.type]]
directory = "security"
name = "Security"
showcontent = true

[[tool.towncrier.type]]
directory = "analysis"
name = "Analyses"
showcontent = true

[[tool.towncrier.type]]
directory = "docs"
name = "Improved Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "maintenance"
name = "Maintenance Changes"
showcontent = true
